Refactor: hash-based keys, segment-centric architecture, metadata package
Major architectural changes:
- Changed Key from []byte to uint64 (hash-based keys for efficiency)
- Introduced metadata package for SegmentRecord/BlobRecord encoding with checksumming
- Dual-level indexing: in-memory skipmap for blobs + Bitcask for segment metadata
- Unified Storage type replacing BlobReader/StorageStrategy abstractions
- Removed context.Context throughout (synchronous operations)
- BlobRecord.Checksum expanded to uint64 with InvalidChecksum sentinel

New metadata package provides:
- AppendBlobRecord/DecodeBlobRecord - individual blob metadata (32 bytes)
- AppendSegmentRecord/DecodeSegmentRecord - segment metadata with CTime and SegmentID
- AppendSegmentRecordWithFooter - adds checksum validation footer
- DecodeSegmentFooter - two-phase decode for efficient footer reading

Performance improvements:
- Bloom filter rebuild: 20-30s → 0.5s (reads from segment records, always current)
- Removed bloom persistence (no stale data issues)
- DirectIO footer writing fixed (proper alignment through write paths)

Bug fixes:
- Fixed eviction logic (was running when disabled)
- Fixed file handle cleanup in closeCurrentSegment (defer pattern prevents double-close)
- Fixed DirectIO footer writes (route through writeDirectIO for alignment)

Deletions: storage.go, segment_reader.go, segment_writer.go, buffered_*.go,
directio_writer.go, paths.go, segment_footer.go (consolidated into segment_storage.go)

+++ blob_writer.go
@@ -16,7 +16,7 @@ type BlobWriter interface {
 
 	// Write writes a blob
 	// checksum is the blob's checksum (0 if not computed)
-	Write(key Key, value []byte, checksum uint32) error
+	Write(key Key, value []byte, checksum uint64) error
 
 	// Pos returns the current write position (for index tracking)
 	// Called after Write() to get position for the last written blob
